♻️ Refactor: kakaomap 관련 타입 위치 수정 (#64)
