+++ src/app/(auth)/eatsplace/_components/PlaceMap.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
 export const PlaceMap = () => {
-  return <div>PlaceMap</div>;
+  //   return   <Map className='h-[150px] w-full rounded-3xl' center={{ lat: y, lng: x }}></Map>
 };
